Update to OAuth 2
Fixes #6

This introduces a few breaking changes, but the good news is that your installations were already broken because of the protocol upgrade. :sweat_smile:

The README has been updated to reflect the necessary changes and steps to go through in obtaining an "Implicit Grant Flow" token. The big difference here is that while the OAuth 1.0a tokens would last forever, these expire after a year. This is slightly better than the normal "Authorization Code Grant Flow" which would issue a token and a refresh that only lasted a few hours at most.

You will definitely need to update:

- Your `FITBIT_CLIENT_ID`
- Your `FITBIT_OAUTH_TOKEN`

To accomplish this, I needed to swap out the underlying library to talk to Fitbit. This means error handling and other conditions likely broke, so please file issues with reporduction steps after this merges if you run into edge cases.