Update fitbit-node, augment test suite
The changes in the underlying `fitbit-node` module has caused an issue
after the configuration options were changed from method arguments to a
configuration hash. The dependencies were not locked down enough and the
minor version was not backwards compatible. This PR adds a test suite in
addition to updating the initializing of the underlying library.