Improve handling of expired tokens
The tokens from Fitbit expire after (at maximum) a year. Prior to this
PR, no error message (or feedback) was returned to the user in these
cases. We will now prompt the user to go through the set up process
again to register the bot (or at least have an administrator do so).